BUG: boxcox constancy check #5112 #10978
If specific lmbda value is passed there is no reason for array to have non-constant values. The problem occurs when trying to fit lmbda variable on a distribution with a single value support.

In case when running a function at inference time (potentially, with a single value), there shouldn't be a restriction on values of x to be non-constant.